fix: refactor pool list by balancer implementation
